feat: Edit user profile

+++ src/Modules/Stakeholders/Explorer.Stakeholders.API/Public/IPersonService.cs
@@ -7,7 +7,7 @@ namespace Explorer.Stakeholders.API.Public
     public interface IPersonService
     {
         Result<PersonDto> Update(PersonDto personDto);
-        Result<PersonDto> GetByUserId(int id);
+        Result<PersonDto> GetByUserId(long id);
         Result<PagedResult<PersonDto>> GetPaged(int page, int pageSize);
     }
 }
